Removed and moved old routines
Kept:
* select_energy_band
* stack_1D_spectrum

Removed:
* sum_diffraction : replaced by stitching.stitch (tested)
* calibrate_detector : replaced by calibration.compute_tth_per_step (tested)
